+++ src/main/java/net/offbeatpioneer/retroengine/core/util/BitmapHelper.java
@@ -7,19 +7,26 @@ import android.graphics.Matrix;
  * Various helper methods for bitmap operations
  *
  * @author Dominik Grzelak
- * @since 01.04.2017
+ * @since 2017-04-01
  */
 public class BitmapHelper {
 
-    public static Bitmap scaleToFit(Bitmap background, float scaleWidth, float scaleHeight) {
-        if (background == null) {
+    /**
+     * Scales a bitmap to fit a specified width and size
+     *
+     * @param bitmap      source bitmap
+     * @param scaleWidth  new width of the bitmap
+     * @param scaleHeight new height of the bitmap
+     * @return new scaled bitmap or null if bitmap was null
+     */
+    public static Bitmap scaleToFit(Bitmap bitmap, float scaleWidth, float scaleHeight) {
+        if (bitmap == null) {
             return null;
         }
 
         Matrix matrix = new Matrix();
         matrix.postScale(scaleWidth, scaleHeight);
 
-        background = Bitmap.createBitmap(background, 0, 0, background.getWidth(), background.getHeight(), matrix, false);
-        return background;
+        return Bitmap.createBitmap(bitmap, 0, 0, bitmap.getWidth(), bitmap.getHeight(), matrix, false);
     }
 }